refactor(auth): deduplicate SLM login endpoints (#1922)
Merge modern features from /api/slm-auth/login into /api/auth/login
and remove the duplicate endpoint. The canonical /api/auth/login now
uses UserService (supports email login), MFA challenges, X-Forwarded-For
IP detection, and standardized audit logging.
