+++ autobot-slm-backend/api/__init__.py
@@ -36,7 +36,6 @@ from .services import fleet_router as fleet_services_router
 from .services import router as services_router
 from .settings import router as settings_router
 from .setup_wizard import router as setup_wizard_router
-from .slm_auth import router as slm_auth_router
 from .slm_users import router as slm_users_router
 from .sso import router as sso_router
 from .sso_auth import router as sso_auth_router
@@ -80,7 +79,6 @@ __all__ = [
     "npu_router",
     "infrastructure_router",
     "slm_users_router",
-    "slm_auth_router",
     "autobot_users_router",
     "autobot_teams_router",
     "sso_router",
